refactor(e2e): Split e2e test files into focused modules
Split large e2e test files for better organization and faster
targeted test runs.

- 20-markdown-editor.spec.ts (2,019 lines): Split into 4 focused
  specs — toolbar, input, paste, draft
- helpers.ts (760 lines): Split into helpers/ directory with 5
  focused sub-modules — server, toast, ui, api, plan-mode
- Update all 46 consumer files to import from specific sub-modules
