+++ frontend/tests/e2e/07-org-management.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test } from './fixtures'
-import { loginViaUI } from './helpers'
+import { loginViaUI } from './helpers/ui'
 
 test.describe('Organization Management', () => {
   test('should navigate to org management page', async ({ page }) => {
