+++ frontend/tests/e2e/10-user-preferences.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test } from './fixtures'
-import { loginViaUI } from './helpers'
+import { loginViaUI } from './helpers/ui'
 
 test.describe('User Preferences', () => {
   test('should navigate to settings page', async ({ page }) => {
