+++ frontend/tests/e2e/20-markdown-editor-input.spec.ts
@@ -1,122 +1,4 @@
 import { expect, test } from './fixtures'
-import { enterAndExitPlanMode, openAgentViaUI } from './helpers'
-
-test.describe('Markdown Editor Toolbar', () => {
-  test('should show active state on bold button when text is bold', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-
-    // Click bold button
-    const boldBtn = page.locator('[data-testid="toolbar-bold"]')
-    await boldBtn.click()
-
-    // Bold button should have the active class
-    await expect(boldBtn).toHaveClass(/IconButton_active/)
-
-    // Type some text — it should be bold
-    await editor.click()
-    await page.keyboard.type('bold text')
-    await expect(editor.locator('strong')).toHaveText('bold text')
-
-    // Bold button should still be active
-    await expect(boldBtn).toHaveClass(/IconButton_active/)
-
-    // Toggle bold off
-    await boldBtn.click()
-    await expect(boldBtn).not.toHaveClass(/IconButton_active/)
-  })
-
-  test('should select heading level from dropdown', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-
-    // Click in editor first
-    await editor.click()
-    await page.keyboard.type('Some text')
-
-    // Open heading dropdown
-    const headingBtn = page.locator('[data-testid="toolbar-heading"]')
-    await headingBtn.click()
-
-    // Menu should appear
-    const menu = page.locator('[data-testid="heading-menu"]')
-    await expect(menu).toBeVisible()
-
-    // Select Heading 2
-    await page.locator('[data-testid="heading-2"]').click()
-
-    // An h2 element should be in the editor
-    await expect(editor.locator('h2')).toBeVisible()
-
-    // Heading button should now be active
-    await expect(headingBtn).toHaveClass(/IconButton_active/)
-
-    // Convert back to paragraph
-    await headingBtn.click()
-    await expect(menu).toBeVisible()
-    await page.locator('[data-testid="heading-paragraph"]').click()
-
-    // h2 should be gone
-    await expect(editor.locator('h2')).not.toBeVisible()
-    await expect(headingBtn).not.toHaveClass(/IconButton_active/)
-  })
-
-  test('should insert code block when clicking code block button', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-
-    await editor.click()
-
-    // Click code block button
-    const codeBlockBtn = page.locator('[data-testid="toolbar-codeblock"]')
-    await codeBlockBtn.click()
-
-    // A pre element should appear in the editor
-    await expect(editor.locator('pre')).toBeVisible()
-
-    // Code block button should be active
-    await expect(codeBlockBtn).toHaveClass(/IconButton_active/)
-  })
-
-  test('should grow editor beyond old 120px limit', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-
-    // Default mode is Cmd+Enter-to-send, so Enter creates newlines
-    await editor.click()
-
-    // Type many lines to exceed the old 120px limit
-    for (let i = 0; i < 15; i++) {
-      await page.keyboard.type(`Line ${i + 1}`, { delay: 100 })
-      await page.keyboard.press('Enter')
-    }
-
-    // The editor wrapper should have grown (capped at 75% of container)
-    const wrapper = page.locator('[data-testid="chat-editor"]')
-    const height = await wrapper.evaluate(el => el.getBoundingClientRect().height)
-    expect(height).toBeGreaterThan(60)
-  })
-
-  test('should use --mono-font-family CSS variable for code elements', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-
-    await editor.click()
-    await page.keyboard.type('some code')
-
-    // Select all text, then apply inline code via toolbar
-    await page.keyboard.press('Meta+a')
-    const codeBtn = page.locator('[data-testid="toolbar-code"]')
-    await codeBtn.click()
-
-    // Check that the code element's computed font-family includes the fallback
-    const fontFamily = await editor.locator('code').evaluate(
-      el => window.getComputedStyle(el).fontFamily,
-    )
-    // Should contain at least one of the expected monospace fonts
-    expect(fontFamily).toMatch(/HackNerdFont|Menlo|Monaco|Courier New|monospace/)
-  })
-})
 
 test.describe('Tab and Shift+Tab behavior', () => {
   test('Tab indents bullet list item', async ({ page, authenticatedWorkspace }) => {
@@ -420,85 +302,6 @@ test.describe('Tab and Shift+Tab behavior', () => {
   })
 })
 
-test.describe('Draft Persistence', () => {
-  test('draft survives page reload', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-
-    // Type some text
-    await editor.click()
-    await page.keyboard.type('draft text to preserve', { delay: 100 })
-
-    // Wait for debounced save (500ms)
-    await page.waitForTimeout(700)
-
-    // Reload the page
-    await page.reload()
-
-    // Wait for the editor to appear again
-    await expect(page.locator('[data-testid="chat-editor"] .ProseMirror')).toBeVisible()
-
-    // The draft text should be restored
-    const restoredEditor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(restoredEditor).toContainText('draft text to preserve')
-  })
-
-  test('draft is cleared after sending', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-
-    // Type and wait for save
-    await editor.click()
-    await page.keyboard.type('message to send', { delay: 100 })
-    await page.waitForTimeout(700)
-
-    // Send the message (default mode is Cmd+Enter-to-send)
-    await page.keyboard.press('Meta+Enter')
-
-    // Wait a moment for the clear to take effect
-    await page.waitForTimeout(200)
-
-    // Reload
-    await page.reload()
-    await expect(page.locator('[data-testid="chat-editor"] .ProseMirror')).toBeVisible()
-
-    // Editor should be empty (draft was cleared on send)
-    const restoredEditor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    const text = await restoredEditor.textContent()
-    expect(text?.trim()).toBe('')
-  })
-
-  test('each agent tab has isolated draft', async ({ page, authenticatedWorkspace }) => {
-    // Type in the first agent
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-    await editor.click()
-    await page.keyboard.type('first agent text', { delay: 100 })
-    await page.waitForTimeout(700)
-
-    // Open a second agent tab
-    await openAgentViaUI(page)
-
-    // The second agent should have an empty editor (wait for draft swap)
-    const editor2 = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor2).toBeVisible()
-    await page.waitForTimeout(500)
-    await expect(editor2).not.toContainText('first agent text')
-
-    // Type in the second agent
-    await editor2.click()
-    await page.keyboard.type('second agent text', { delay: 100 })
-    await page.waitForTimeout(700)
-
-    // Switch back to the first agent tab
-    const agentTabs = page.locator('[data-testid="tab"][data-tab-type="agent"]')
-    await agentTabs.first().click()
-
-    // First agent should still have its text
-    await expect(page.locator('[data-testid="chat-editor"] .ProseMirror')).toContainText('first agent text')
-  })
-})
-
 test.describe('Horizontal Rule Input Rule', () => {
   test('typing --- creates a horizontal rule', async ({ page, authenticatedWorkspace }) => {
     // Default mode is Cmd+Enter-to-send, so Enter creates newlines
@@ -558,91 +361,6 @@ test.describe('List Input Rules After Soft Break', () => {
 
     // An ordered list should be created
     await expect(editor.locator('ol')).toBeVisible()
-  })
-})
-
-test.describe('Toggleable List Buttons', () => {
-  test('bullet list button toggles on and off', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-    await editor.click()
-
-    const bulletBtn = page.locator('[data-testid="toolbar-bullet-list"]')
-
-    // Click bullet list → should activate
-    await bulletBtn.click()
-    await expect(bulletBtn).toHaveClass(/IconButton_active/)
-    await expect(editor.locator('ul')).toBeVisible()
-
-    // Click again → should deactivate (convert back to paragraph)
-    await bulletBtn.click()
-    await expect(bulletBtn).not.toHaveClass(/IconButton_active/)
-    await expect(editor.locator('ul')).not.toBeVisible()
-  })
-
-  test('switching from bullet to ordered list', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-    await editor.click()
-    await page.keyboard.type('list item', { delay: 100 })
-
-    const bulletBtn = page.locator('[data-testid="toolbar-bullet-list"]')
-    const orderedBtn = page.locator('[data-testid="toolbar-ordered-list"]')
-
-    // Create bullet list
-    await bulletBtn.click()
-    await expect(editor.locator('ul')).toBeVisible()
-    await expect(bulletBtn).toHaveClass(/IconButton_active/)
-
-    // Switch to ordered list
-    await orderedBtn.click()
-    await expect(editor.locator('ol')).toBeVisible()
-    await expect(editor.locator('ul')).not.toBeVisible()
-    await expect(orderedBtn).toHaveClass(/IconButton_active/)
-    await expect(bulletBtn).not.toHaveClass(/IconButton_active/)
-  })
-
-  test('task list button wraps existing text into task list', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-    await editor.click()
-    await page.keyboard.type('my text', { delay: 100 })
-
-    const taskBtn = page.locator('[data-testid="toolbar-tasklist"]')
-
-    // Click task list button — should wrap the existing text into a task list item
-    await taskBtn.click()
-    await expect(taskBtn).toHaveClass(/IconButton_active/)
-
-    // The task list item should contain the text we typed
-    const taskItem = editor.locator('li[data-checked]')
-    await expect(taskItem).toBeVisible()
-    await expect(taskItem).toContainText('my text')
-
-    // There should be no separate paragraph outside the list
-    // (the text should not remain as a standalone paragraph)
-    const topLevelParagraphs = editor.locator(':scope > p')
-    await expect(topLevelParagraphs).toHaveCount(0)
-  })
-
-  test('switching from bullet to task list', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-    await editor.click()
-    await page.keyboard.type('task item', { delay: 100 })
-
-    const bulletBtn = page.locator('[data-testid="toolbar-bullet-list"]')
-    const taskBtn = page.locator('[data-testid="toolbar-tasklist"]')
-
-    // Create bullet list
-    await bulletBtn.click()
-    await expect(editor.locator('ul')).toBeVisible()
-
-    // Switch to task list
-    await taskBtn.click()
-    await expect(editor.locator('li[data-checked]')).toBeVisible()
-    await expect(taskBtn).toHaveClass(/IconButton_active/)
-    await expect(bulletBtn).not.toHaveClass(/IconButton_active/)
   })
 })
 
@@ -1210,30 +928,6 @@ test.describe('Horizontal Rule Backspace Revert', () => {
   })
 })
 
-test.describe('Markdown Paste', () => {
-  test('pasting markdown list text creates a bullet list', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-    await editor.click()
-
-    // Paste markdown list content via clipboard API
-    await page.evaluate(() => {
-      const data = new DataTransfer()
-      data.setData('text/plain', '- foo\n- bar\n- baz')
-      const event = new ClipboardEvent('paste', { clipboardData: data, bubbles: true, cancelable: true })
-      document.querySelector('.ProseMirror')!.dispatchEvent(event)
-    })
-
-    // A bullet list should be created with the items
-    await expect(editor.locator('ul')).toBeVisible()
-    const items = editor.locator('ul > li')
-    await expect(items).toHaveCount(3)
-    await expect(items.nth(0)).toContainText('foo')
-    await expect(items.nth(1)).toContainText('bar')
-    await expect(items.nth(2)).toContainText('baz')
-  })
-})
-
 test.describe('Code Block Input Rule', () => {
   test('typing ``` creates a code block', async ({ page, authenticatedWorkspace }) => {
     const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
@@ -1281,90 +975,6 @@ test.describe('Link Input Rule', () => {
     await expect(link).toBeVisible()
     await expect(link).toHaveText('click here')
     await expect(link).toHaveAttribute('href', 'https://example.com')
-  })
-})
-
-test.describe('Clipboard Copy/Paste', () => {
-  test('copy and paste preserves markdown structure', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-    await editor.click()
-
-    // Create bold text
-    const boldBtn = page.locator('[data-testid="toolbar-bold"]')
-    await boldBtn.click()
-    await page.keyboard.type('bold text', { delay: 100 })
-    await boldBtn.click()
-
-    // Verify bold text exists
-    await expect(editor.locator('strong')).toHaveText('bold text')
-
-    // Step 1: Verify that the copy serializer produces markdown with
-    // inline formatting (our clipboardTextSerializer override).
-    // Capture clipboard text by intercepting DataTransfer.setData
-    // during a copy triggered via execCommand.
-    await page.keyboard.press('Meta+a')
-    const clipboardText = await page.evaluate(() => {
-      let captured = ''
-      const origSetData = DataTransfer.prototype.setData
-      DataTransfer.prototype.setData = function (format: string, value: string) {
-        if (format === 'text/plain')
-          captured = value
-        return origSetData.call(this, format, value)
-      }
-      document.execCommand('copy')
-      DataTransfer.prototype.setData = origSetData
-      return captured
-    })
-    // clipboardTextSerializer should produce markdown with bold markers
-    expect(clipboardText).toContain('**')
-
-    // Step 2: Verify that pasting markdown with inline formatting
-    // restores the marks (our enhanced paste plugin). Clear the editor
-    // first, then paste the captured markdown via synthetic paste event.
-    await page.keyboard.press('Meta+a')
-    await page.keyboard.press('Backspace')
-    await expect(editor.locator('strong')).toHaveCount(0)
-
-    await page.evaluate((text) => {
-      const dt = new DataTransfer()
-      dt.setData('text/plain', text)
-      const event = new ClipboardEvent('paste', { clipboardData: dt, bubbles: true, cancelable: true })
-      document.querySelector('.ProseMirror')!.dispatchEvent(event)
-    }, clipboardText)
-
-    // The pasted markdown should produce bold text
-    await expect(editor.locator('strong')).toHaveText('bold text')
-  })
-})
-
-test.describe('Code Language Label', () => {
-  test('clicking language label opens popover and selecting language updates it', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-    await editor.click()
-
-    // Create a code block
-    const codeBlockBtn = page.locator('[data-testid="toolbar-codeblock"]')
-    await codeBlockBtn.click()
-
-    // Language label should show "plaintext" by default
-    const langLabel = editor.locator('.code-lang-label')
-    await expect(langLabel).toHaveText('plaintext')
-
-    // Click the language label
-    await langLabel.click()
-
-    // The language input should appear
-    const langInput = page.locator('[data-testid="code-lang-input"]')
-    await expect(langInput).toBeVisible()
-
-    // Type and select a language
-    await langInput.fill('python')
-    await page.keyboard.press('Enter')
-
-    // The label should now show "python"
-    await expect(langLabel).toHaveText('python')
   })
 })
 
@@ -1637,110 +1247,6 @@ test.describe('ArrowDown Code Block Scroll', () => {
       expect(paraBox.y).toBeGreaterThanOrEqual(wrapperBox.y)
       expect(paraBox.y).toBeLessThanOrEqual(wrapperBox.y + wrapperBox.height)
     }
-  })
-})
-
-test.describe('Send Feedback Button Labels', () => {
-  test('ExitPlanMode banner shows Reject when editor is empty, Send Feedback when typing', async ({ page, authenticatedWorkspace }) => {
-    // Enter plan mode, write a dummy plan, and exit
-    const banner = await enterAndExitPlanMode(page)
-    await expect(banner.getByText('Plan Ready for Review')).toBeVisible()
-
-    const rejectBtn = page.locator('[data-testid="plan-reject-btn"]')
-
-    // With empty editor, button should say "Reject"
-    await expect(rejectBtn).toHaveText('Reject')
-
-    // Type feedback into the editor — button should switch to "Send Feedback"
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await editor.click()
-    await page.keyboard.type('Please reconsider this approach')
-    await expect(rejectBtn).toHaveText('Send Feedback')
-  })
-})
-
-test.describe('Paste Into Code Context', () => {
-  test('paste fenced code block into code_block strips delimiters', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-    await editor.click()
-
-    // Create a code block
-    const codeBlockBtn = page.locator('[data-testid="toolbar-codeblock"]')
-    await codeBlockBtn.click()
-    await expect(editor.locator('pre')).toBeVisible()
-
-    // Set clipboard to a fenced code block and paste
-    await page.evaluate(() => {
-      const dt = new DataTransfer()
-      dt.setData('text/plain', '```python\nprint("hello")\n```')
-      const event = new ClipboardEvent('paste', { clipboardData: dt, bubbles: true, cancelable: true })
-      document.querySelector('.ProseMirror')!.dispatchEvent(event)
-    })
-    await page.waitForTimeout(200)
-
-    // Should have stripped the fence delimiters
-    const codeText = await editor.locator('pre code').evaluate((el) => {
-      const clone = el.cloneNode(true) as HTMLElement
-      clone.querySelectorAll('.code-lang-label').forEach(s => s.remove())
-      return clone.textContent
-    })
-    expect(codeText).toBe('print("hello")')
-    expect(codeText).not.toContain('```')
-  })
-
-  test('paste inline code into code_block strips backticks', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-    await editor.click()
-
-    // Create a code block
-    const codeBlockBtn = page.locator('[data-testid="toolbar-codeblock"]')
-    await codeBlockBtn.click()
-    await expect(editor.locator('pre')).toBeVisible()
-
-    // Paste inline code
-    await page.evaluate(() => {
-      const dt = new DataTransfer()
-      dt.setData('text/plain', '`myVariable`')
-      const event = new ClipboardEvent('paste', { clipboardData: dt, bubbles: true, cancelable: true })
-      document.querySelector('.ProseMirror')!.dispatchEvent(event)
-    })
-    await page.waitForTimeout(200)
-
-    const codeText = await editor.locator('pre code').evaluate((el) => {
-      const clone = el.cloneNode(true) as HTMLElement
-      clone.querySelectorAll('.code-lang-label').forEach(s => s.remove())
-      return clone.textContent
-    })
-    expect(codeText).toBe('myVariable')
-  })
-
-  test('paste plain text into code_block is unchanged', async ({ page, authenticatedWorkspace }) => {
-    const editor = page.locator('[data-testid="chat-editor"] .ProseMirror')
-    await expect(editor).toBeVisible()
-    await editor.click()
-
-    // Create a code block
-    const codeBlockBtn = page.locator('[data-testid="toolbar-codeblock"]')
-    await codeBlockBtn.click()
-    await expect(editor.locator('pre')).toBeVisible()
-
-    // Paste plain text (no backticks)
-    await page.evaluate(() => {
-      const dt = new DataTransfer()
-      dt.setData('text/plain', 'just plain text')
-      const event = new ClipboardEvent('paste', { clipboardData: dt, bubbles: true, cancelable: true })
-      document.querySelector('.ProseMirror')!.dispatchEvent(event)
-    })
-    await page.waitForTimeout(200)
-
-    const codeText = await editor.locator('pre code').evaluate((el) => {
-      const clone = el.cloneNode(true) as HTMLElement
-      clone.querySelectorAll('.code-lang-label').forEach(s => s.remove())
-      return clone.textContent
-    })
-    expect(codeText).toBe('just plain text')
   })
 })
