+++ frontend/tests/e2e/25-no-worker-settings.spec.ts
@@ -1,4 +1,5 @@
-import { createWorkspaceViaAPI, deleteWorkspaceViaAPI, loginViaToken, waitForWorkspaceReady } from './helpers'
+import { createWorkspaceViaAPI, deleteWorkspaceViaAPI } from './helpers/api'
+import { loginViaToken, waitForWorkspaceReady } from './helpers/ui'
 import { ensureWorkerOnline, expect, restartHub, stopHub, stopWorker, processTest as test } from './process-control-fixtures'
 
 test.describe('Settings and /clear without Worker', () => {
