+++ frontend/tests/e2e/62-directory-tree.spec.ts
@@ -1,11 +1,7 @@
 import process from 'node:process'
 import { expect, test } from './fixtures'
-import {
-  createWorkspaceViaAPI,
-  deleteWorkspaceViaAPI,
-  loginViaToken,
-  waitForWorkspaceReady,
-} from './helpers'
+import { createWorkspaceViaAPI, deleteWorkspaceViaAPI } from './helpers/api'
+import { loginViaToken, waitForWorkspaceReady } from './helpers/ui'
 
 test.describe('DirectoryTree', () => {
   test('root directory is visible and collapsible', async ({ page, leapmuxServer }) => {
