+++ frontend/tests/e2e/fixtures.ts
@@ -10,18 +10,14 @@ import {
   createWorkspaceViaAPI,
   deleteWorkspaceViaAPI,
   enableSignupViaAPI,
-  findFreePort,
   getAdminOrgId,
-  getGlobalState,
-  getRecordedToasts,
   getWorkerId,
-  installToastRecorder,
   loginViaAPI,
-  loginViaToken,
   signUpViaAPI,
-  waitForServer,
-  waitForWorkspaceReady,
-} from './helpers'
+} from './helpers/api'
+import { findFreePort, getGlobalState, waitForServer } from './helpers/server'
+import { getRecordedToasts, installToastRecorder } from './helpers/toast'
+import { loginViaToken, waitForWorkspaceReady } from './helpers/ui'
 
 export interface ServerInfo {
   hubUrl: string
